+++ Wrappers/Python/cil/optimisation/operators/GradientOperator.py
@@ -20,6 +20,7 @@ from cil.optimisation.operators import FiniteDifferenceOperator
 from cil.framework import BlockGeometry
 import warnings
 from cil.utilities.multiprocessing import NUM_THREADS
+from cil.framework import ImageGeometry
 import numpy as np
 
 NEUMANN = 'Neumann'
@@ -79,13 +80,22 @@ class GradientOperator(LinearOperator):
 
         # Default correlation for the gradient coupling
         correlation = kwargs.get('correlation',CORRELATION_SPACE)
+
+        # Add attributes for SIRF data where there is no CIL geometry
+        if not isinstance(domain_geometry, ImageGeometry):
+            domain_geometry.channels = 1
+            domain_geometry.dimension_labels = [None]*len(domain_geometry.shape)        
                        
         # Space correlation on multichannel data call numpy backend
         if correlation == CORRELATION_SPACE and domain_geometry.channels > 1:
             #numpy implementation only for now
             backend = NUMPY
             warnings.warn("Warning: correlation='Space' on multi-channel dataset will use `numpy` backend")
-                     
+
+        # Complex data will use numpy backend
+        if domain_geometry.dtype in [np.complex, np.complex64]:
+            backend = NUMPY
+            warnings.warn("Warning: Complex geometries will use `numpy` backend")
         
         if method != 'forward':
             backend = NUMPY
@@ -406,3 +416,4 @@ class Gradient_C(LinearOperator):
                 
         if return_val is True:
             return out        
+