+++ Wrappers/Python/test/test_SIRF.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 #   This work is part of the Core Imaging Library (CIL) developed by CCPi 
 #   (Collaborative Computational Project in Tomographic Imaging), with 
 #   substantial contributions by UKRI-STFC and University of Manchester.
@@ -15,24 +16,141 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import numpy
+import numpy as np
 from numpy.linalg import norm
 import os
-import sys
 import shutil
 import unittest
 from cil.framework import BlockDataContainer
+from cil.optimisation.operators import GradientOperator, LinearOperator
+
 try:
     import sirf.STIR as pet
-    from sirf.Utilities import  examples_data_path
+    import sirf.Gadgetron as mr
+    from sirf.Utilities import examples_data_path
     has_sirf = True
 except ImportError as ie:
     has_sirf = False
 
+class GradientSIRF(object):
+    
+    @unittest.skipUnless(has_sirf, "Skipping as SIRF is not available")
+    def test_Gradient(self):
+
+        #######################################
+        ##### Test Gradient numpy backend #####
+        #######################################
+
+        Grad_numpy = GradientOperator(self.image1, backend='numpy')
+
+        res1 = Grad_numpy.direct(self.image1)         
+        res2 = Grad_numpy.range_geometry().allocate()
+        Grad_numpy.direct(self.image1, out=res2)
+
+        self.assertTrue(isinstance(res1,BlockDataContainer))
+        self.assertTrue(isinstance(res2,BlockDataContainer))
+
+        for i in range(len(res1)):
+        
+            if isinstance(self.image1, pet.ImageData):
+                self.assertTrue(isinstance(res1[i], pet.ImageData))
+                self.assertTrue(isinstance(res2[i], pet.ImageData)) 
+            else:
+                self.assertTrue(isinstance(res1[i], mr.ImageData))
+                self.assertTrue(isinstance(res2[i], mr.ImageData))                                 
+            # test direct with and without out
+            np.testing.assert_array_almost_equal(res1[i].as_array(), res2[i].as_array())                         
+
+        # test adjoint with and without out
+        res3 = Grad_numpy.adjoint(res1)
+        res4 = Grad_numpy.domain_geometry().allocate()
+        Grad_numpy.adjoint(res2, out=res4)
+        np.testing.assert_array_almost_equal(res3.as_array(), res4.as_array()) 
+
+        # test dot_test
+        self.assertTrue(LinearOperator.dot_test(Grad_numpy, decimal=3))
+
+        # test shape of output of direct
+        
+        # check in the case of pseudo 2D data, e.g., (1, 155, 155)
+        if 1 in self.image1.shape:
+            self.assertEqual(res1.shape, (2,1))
+        else:
+            self.assertEqual(res1.shape, (3,1))            
+
+        ########################################
+        ##### Test Gradient c backend  #####
+        ########################################
+        Grad_c = GradientOperator(self.image1, backend='c')
+
+        # test direct with and without out
+        res5 = Grad_c.direct(self.image1) 
+        res6 = Grad_c.range_geometry().allocate()*0.
+        Grad_c.direct(self.image1, out=res6)
+
+        for i in range(len(res5)):
+            np.testing.assert_array_almost_equal(res5[i].as_array(), res6[i].as_array())
+
+            # compare c vs numpy gradient backends 
+            np.testing.assert_array_almost_equal(res6[i].as_array(), res2[i].as_array())
+
+
+        # test dot_test
+        self.assertTrue(LinearOperator.dot_test(Grad_c, decimal=3))
+
+        # test adjoint
+        res7 = Grad_c.adjoint(res5) 
+        res8 = Grad_c.domain_geometry().allocate()*0.
+        Grad_c.adjoint(res5, out=res8)
+        np.testing.assert_array_almost_equal(res7.as_array(), res8.as_array()) 
+
+
+
+class TestGradientPET_2D(unittest.TestCase, GradientSIRF):  
+
+    def setUp(self):
+
+        if has_sirf:
+            self.image1 = pet.ImageData(os.path.join(
+                examples_data_path('PET'),'thorax_single_slice','emission.hv')
+                )
+
+    def tearDown(self):
+        pass    
+
+class TestGradientPET_3D(unittest.TestCase, GradientSIRF):  
+
+    def setUp(self):
+
+        if has_sirf:
+            self.image1 = pet.ImageData(os.path.join(
+                examples_data_path('PET'),'brain','emission.hv')
+                )
+
+    def tearDown(self):
+        pass  
+
+class TestGradientMR_2D(unittest.TestCase, GradientSIRF):  
+
+    def setUp(self):
+
+        if has_sirf:
+            acq_data = mr.AcquisitionData(os.path.join
+                (examples_data_path('MR'),'simulated_MR_2D_cartesian.h5')
+            )
+            preprocessed_data = mr.preprocess_acquisition_data(acq_data)
+            recon = mr.FullySampledReconstructor()
+            recon.set_input(preprocessed_data)
+            recon.process()
+            self.image1 = recon.get_output()
+        
+    def tearDown(self):
+        pass      
 
 class TestSIRFCILIntegration(unittest.TestCase):
     
     def setUp(self):
+
         if has_sirf:
             os.chdir(examples_data_path('PET'))
             # Copy files to a working folder and change directory to where these files are.
@@ -56,12 +174,11 @@ class TestSIRFCILIntegration(unittest.TestCase):
         image2 = pet.ImageData('emission.hv')
         image1.fill(1.)
         image2.fill(2.)
-        print (image1.shape, image2.shape)
         
         tmp = image1.divide(1.)
-        # numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
+        np.testing.assert_array_equal(image1.as_array(), tmp.as_array())
         tmp = image2.divide(1.)
-        # numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
+        np.testing.assert_array_equal(image2.as_array(), tmp.as_array())
         
 
         # image.fill(1.)
@@ -77,12 +194,11 @@ class TestSIRFCILIntegration(unittest.TestCase):
         image2 = pet.ImageData('emission.hv')
         image1.fill(1.)
         image2.fill(2.)
-        print (image1.shape, image2.shape)
         
         tmp = image1.multiply(1.)
-        numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
+        np.testing.assert_array_equal(image1.as_array(), tmp.as_array())
         tmp = image2.multiply(1.)
-        numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
+        np.testing.assert_array_equal(image2.as_array(), tmp.as_array())
         
 
         # image.fill(1.)
@@ -98,12 +214,11 @@ class TestSIRFCILIntegration(unittest.TestCase):
         image2 = pet.ImageData('emission.hv')
         image1.fill(0)
         image2.fill(1)
-        print (image1.shape, image2.shape)
         
         tmp = image1.add(1.)
-        numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
+        np.testing.assert_array_equal(image2.as_array(), tmp.as_array())
         tmp = image2.subtract(1.)
-        numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
+        np.testing.assert_array_equal(image1.as_array(), tmp.as_array())
         
         bdc = BlockDataContainer(image1, image2)
         bdc1 = bdc.add(1.)
@@ -114,7 +229,7 @@ class TestSIRFCILIntegration(unittest.TestCase):
         bdc = BlockDataContainer(image1, image2)
 
         self.assertBlockDataContainerEqual(bdc , bdc1)
-    
+
     @unittest.skipUnless(has_sirf, "Has SIRF")
     def test_BlockDataContainer_with_SIRF_DataContainer_subtract(self):
         os.chdir(self.cwd)
@@ -122,7 +237,7 @@ class TestSIRFCILIntegration(unittest.TestCase):
         image2 = pet.ImageData('emission.hv')
         image1.fill(2)
         image2.fill(1)
-        print (image1.shape, image2.shape)
+
         
         bdc = BlockDataContainer(image1, image2)
         bdc1 = bdc.subtract(1.)
@@ -133,7 +248,6 @@ class TestSIRFCILIntegration(unittest.TestCase):
         bdc = BlockDataContainer(image1, image2)
 
         self.assertBlockDataContainerEqual(bdc , bdc1)
-
 
     def assertBlockDataContainerEqual(self, container1, container2):
         print ("assert Block Data Container Equal")
@@ -149,4 +263,4 @@ class TestSIRFCILIntegration(unittest.TestCase):
                 self.assertBlockDataContainerEqual(container1.get_item(col),container2.get_item(col))
     
     def assertNumpyArrayEqual(self, first, second):
-        numpy.testing.assert_array_equal(first, second)
+        np.testing.assert_array_equal(first, second)